разделена логика для возврата id и поля для выбора сущности, созданы методы для выполнения Action с возвратом entityId проведен рефакторинг
