+++ MigrationToDRX/Data/Services/FileService.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Text;
 
 namespace MigrationToDRX.Data.Services;
