+++ MigrationToDRX/Data/Services/ODataEDocService.cs
@@ -1,5 +1,6 @@
 using System;
 using Microsoft.Extensions.Logging;
+using MigrationToDRX.Data.Constants;
 using NLog;
 using Simple.OData.Client;
 
@@ -10,18 +11,142 @@ namespace MigrationToDRX.Data.Services;
 /// </summary>
 public class ODataEDocService
 {
+    private readonly OdataClientService _odataClientService;
+    private readonly FileService _fileService;
     private readonly ODataClient _client;
-    private readonly OdataClientService _service;
     private readonly Logger logger = LogManager.GetCurrentClassLogger();
+    
 
     /// <summary>
     /// Конструктор
     /// </summary>
     /// <param name="client"></param>
-    public ODataEDocService(OdataClientService service)
+    public ODataEDocService(OdataClientService odataClientService, FileService fileService)
     {
-        _service = service;
-        _client = _service.GetClient();
+        _odataClientService = odataClientService;
+        _fileService = fileService;
+        _client = _odataClientService.GetClient();
+    }
+
+    /// <summary>
+    /// Создать тело документа
+    /// </summary>
+    public async Task<IDictionary<string, object>?> FindEdocAndSetBodyAsync(long eDocId, string filePath, CancellationToken ct, bool ForceUpdateBody = false)
+    {
+        ct.ThrowIfCancellationRequested();
+
+        var eDoc = await _odataClientService.FindEdocAsync(eDocId, ct);
+
+        if (eDoc == null)
+        {
+            throw new ArgumentException("Не удалось найти документ");
+        }
+
+        var extension = Path.GetExtension(filePath).Replace(".", "");
+
+        eDoc.TryGetValue(OdataPropertyNames.Id, out var docIdObj);
+
+        if (docIdObj == null)
+        {
+            throw new ArgumentException("Не удалось найти Id документа");
+        }
+
+        var docId = Convert.ToInt64(docIdObj);
+
+        var version = GetLastVersion(extension, eDoc);
+        var targetApp = await _odataClientService.FindAssociatedApplication(extension, ct);
+
+        if (targetApp == null)
+        {
+            return null;
+        }
+
+        if (version == null)
+        {
+            version = await _odataClientService.CreateNewVersion(docId, "Первоначальная версия", targetApp, ct);
+        }
+
+        else if (NeedNewVersion(extension, version) || ForceUpdateBody == true)
+        {
+            version = await _odataClientService.CreateNewVersion(docId, Path.GetFileName(filePath), targetApp, ct);
+        }
+
+        byte[] body;
+        try
+        {
+            body = await _fileService.ReadFileEvenIfOpenAsync(filePath);
+
+            if (body.Length == 0)
+            {
+                body = await File.ReadAllBytesAsync(filePath);
+            }
+        }
+        catch (Exception ex)
+        {
+            throw new Exception($"Не удалось прочитать файл: {ex.Message}");
+        }
+
+        await _odataClientService.FillBodyAsync(docId, body, version!, ct);
+
+        return eDoc;
+    }
+
+    /// <summary>
+    /// Проверяет необходимость создания новой версии
+    /// </summary>
+    /// <param name="extension">расширение файла</param>
+    /// <param name="version">верия</param>
+    private bool NeedNewVersion(string extension, IDictionary<string, object> version)
+    {
+        return true;
+
+        // Не удалять!
+        //TODO: Добавить логический оператор для выбра, нужно ли перезатирать текущую версию
+
+        // if (!version.TryGetValue(PropertyNames.AssociatedApplication, out var application))
+        // { return true; }
+
+        // if (application is IDictionary<string, object> currentApp && currentApp.TryGetValue(PropertyNames.Extension, out var currentExtension))
+        // { return currentExtension.ToString() != extension; }
+
+        // return true;
+
+    }
+
+    /// <summary>
+    /// Получает последнюю версию документа
+    /// </summary>
+    /// <param name="extension">расширение файла</param>
+    /// <param name="eDoc">документ</param>
+    /// <returns>последняя версия</returns>
+    private IDictionary<string, object>? GetLastVersion(string extension, IDictionary<string, object> eDoc)
+    {
+        if (eDoc.TryGetValue(OdataPropertyNames.Versions, out var versions))
+        {
+            if (versions == null)
+            {
+                return null;
+            }
+
+            if (versions is IEnumerable<IDictionary<string, object>> versionList)
+            {
+                var version = versionList
+                    .Select(version => new
+                    {
+                        Version = version,
+                        Number = version[OdataPropertyNames.Number] as int? ?? 0 // Преобразование "Number" в int
+                    })
+                    .OrderByDescending(version => version.Number)
+                    .FirstOrDefault();
+
+                return version?.Version ?? null;
+            }
+            else if (versions is IDictionary<string, object> versionDict)
+            {
+                return versionDict;
+            }
+        }
+        return null;
     }
 
     /// <summary>
